renamed get token model
